% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRequest.R
\name{checkRequest}
\alias{checkRequest}
\title{Check if API request returns an error or success.}
\usage{
checkRequest(request)
}
\arguments{
\item{request}{Object returned by GET()}
}
\description{
Stop and print an error message if API call doesn't return 200 status code.
}
\keyword{internal}
