% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructGET.R
\name{constructGET}
\alias{constructGET}
\title{Constructs URL of a full API call.}
\usage{
constructGET(base_url, params)
}
\arguments{
\item{base_url}{string, base url of API.}

\item{params}{list of parameter values.}
}
\value{
url of full API call.
}
\description{
Used internally by searchLocation().
}
\seealso{
\code{\link{searchLocation}}.
}
\keyword{internal}
