% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseResults.R
\name{rmPrefix}
\alias{rmPrefix}
\title{Extracts values from parsed data.}
\usage{
rmPrefix(name, prefix)
}
\arguments{
\item{results}{list of queried and parsed data.}
}
\value{
a data frame.
}
\description{
Used internally by searchLocation().
}
\seealso{
\code{\link{searchLocation}}.
}
\keyword{internal}
