% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchLocation.R
\name{searchLocation}
\alias{searchLocation}
\title{Searches for Airbnb listings by location via its API.}
\usage{
searchLocation(location, api_key = "d306zoyjsyarp7ifhu67rjxn52tv0t20",
  checkin = Sys.Date(), checkout = Sys.Date() + 1, guests = 1,
  min_bathrooms = 1, min_bedrooms = 1, min_beds = 1, price_min = 0,
  price_max = 25, min_num_pics = 5, verbose = F)
}
\arguments{
\item{location}{string, the desired search place, usually a city name. Doesn't 
matter if capitalized or not. For example, Osaka or osaka.}

\item{api_key}{string, your own Airbnb API key. Ours has been provided for 
convenience.}

\item{checkin}{string, check in date of format "yyyy-mm-dd".}

\item{checkout}{string, check out date of format "yyyy-mm-dd".}

\item{guests}{integer, number of guests, default = 1.}

\item{min_bathrooms}{numeric, min number of bathrooms, default = 1.}

\item{min_bedrooms}{integer, min number of bedrooms, default = 1.}

\item{price_min}{integer, min price (USD) per night, default = 0.}

\item{price_max}{integer, max price (USD) per night, default = 25.}

\item{min_num_pics}{integer, min number of pictures, default = 5.}

\item{verbose}{boolean, T/F, indicating whether to print status updates.}
}
\value{
A named list containing various search outputs.
}
